fix: 🧩 修复 flushSync was called from inside a lifecycle method. error
