--- src/extensions/ImageBlock/components/ImageBlockView.tsx
@@ -1,50 +1,0 @@
-import { Node } from '@tiptap/pm/model';
-import { Editor, NodeViewWrapper } from '@tiptap/react';
-import { useCallback, useRef } from 'react';
-
-import { cn } from '@/utils/utils';
-
-interface ImageBlockViewProps {
-  editor: Editor;
-  getPos: () => number | undefined;
-  node: Node;
-  updateAttributes: (attrs: Record<string, string>) => void;
-}
-
-export const ImageBlockView = (props: ImageBlockViewProps) => {
-  const { editor, getPos, node } = props as ImageBlockViewProps & {
-    node: Node & {
-      attrs: {
-        src: string;
-      };
-    };
-  };
-  const imageWrapperRef = useRef<HTMLDivElement>(null);
-  const { src } = node.attrs;
-
-  const wrapperClassName = cn(
-    node.attrs.align === 'left' ? 'ml-0' : 'ml-auto',
-    node.attrs.align === 'right' ? 'mr-0' : 'mr-auto',
-    node.attrs.align === 'center' && 'mx-auto',
-  );
-
-  const onClick = useCallback(() => {
-    const pos = getPos();
-
-    if (pos !== undefined) {
-      editor.commands.setNodeSelection(pos);
-    }
-  }, [getPos, editor.commands]);
-
-  return (
-    <NodeViewWrapper>
-      <div className={wrapperClassName} style={{ width: node.attrs.width }} data-drag-handle>
-        <div contentEditable={false} ref={imageWrapperRef}>
-          <img className="block" src={src} alt="" onClick={onClick} />
-        </div>
-      </div>
-    </NodeViewWrapper>
-  );
-};
-
-export default ImageBlockView;
